Parallel data load to ES and mySQL

+++ elasticsearch/snapsMappings.js
@@ -31,7 +31,7 @@ client.indices.create({
                     "type": "date"
                 },
                 "cameraId": {
-                    "type": "short"
+                    "type": "integer"
                 },
                 "LPR": {
                     "type": "keyword"
